A system on a module (SOM) is a Board level circuit that
A system on a module (SOM) is a Board level circuit that integrates a system function in a single module. Shenzhen Graperain Technology Co., Ltd. have spend many years in Samsung and Qualcomm chipset system on module, such as Samsung S5PV210, Samsung Exynos4412, Samsung S5P4418, S5P6818, and Qualcom MSM8916. The engineers design the schematic diagram and product of these system on modules. https://www.graperain.com/ARM-Embedded-System-On-Module/

System-on-Module is the MINI PC's core functions packaged in an electronic board packaging. Most of the system-on-module integrates the CPU, memory devices and pins, through the pin and supporting the bottom plate connected together to achieve a field of system chips. People often such a system called the microcontroller, or an embedded development platform. Because the system-on-module integrates the core of the common functions, so it has a c system-on-module can customize the versatility of the different floor, which greatly improves the development efficiency of the microcontroller. Because the system-on-module as a separate module separated, it also reduces the development of the difficulty, an increase of system stability and maintainability. https://www.graperain.com

These system on module could be used in products directly, just design a carrier board and pin out its functions. The company could provide this design service for the carrier board efficiently from the engineers who designed the SOM. Now these Samsung SOM and Qualcomm SOM have been proved well in market for last two years. And the clients have used them in POS, PDA, PND, smart home, phone, learning machine, game machine and other kinds of industrial control filed. https://www.graperain.com/ARM-Single-Board-Computer/
https://www.graperain.com/Embedded-Hardware-Design/